big time reorganization
lets make it easier to sift thru this code
