+++ .gitignore
@@ -1,1 +1,2 @@
 wrfout*
+__pycache__
